+++ crates/renderer/src/lib.rs
@@ -3,14 +3,14 @@ use freya_native_core::NodeId;
 use std::sync::{Arc, Mutex};
 
 pub use config::WindowConfig;
-pub use window::DesktopRenderer;
+pub use renderer::DesktopRenderer;
 
 mod accessibility;
 mod app;
 mod config;
 mod elements;
+mod render;
 mod renderer;
-mod window;
 mod winit_waker;
 mod wireframe;
